requirements.txt
python-telegram-bot==21.8
docker==7.1.0
python-dotenv==1.0.1
psutil==6.0.0
